+++ objects/src/assets/nonfungible.rs
@@ -255,12 +255,12 @@ mod tests {
     use assert_matches::assert_matches;
 
     use super::*;
-    use crate::accounts::{
-        account_id::testing::{
+    use crate::{
+        accounts::AccountId,
+        testing::account_id::{
             ACCOUNT_ID_FUNGIBLE_FAUCET_OFF_CHAIN, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_OFF_CHAIN,
             ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1,
         },
-        AccountId,
     };
 
     #[test]
